requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
asteroid-filterbanks==0.4.0
attrs==24.3.0
audioread==3.0.1
azure-cognitiveservices-speech==1.41.1
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
Django==5.1.4
djangorestframework==3.15.2
docopt==0.6.2
einops==0.8.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.27.0
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==2.10
Jinja2==3.1.5
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
propcache==0.2.1
protobuf==5.29.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.0
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
textblob==0.18.0.post0
threadpoolctl==3.5.0
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
